pyproject.toml
[project]
name = "fastapi-mvc-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.116.1",
    "jinja2>=3.1.6",
    "motor>=3.7.1",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
